# SolucionHyJ_20191025 (Fase 1)/WS_HyJ/wwwroot/swagger.yaml
swagger: "2.0"
info:
  version: 1.0
  title: REST API - HyJ
  description: Lista de servicios API para la solucion de HyJ.
  termsOfService: ''
  contact:
    name: Diana Rolando
    url: ''
    email: drolando@gmail.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
paths:
...
  /DEV/api/v1/Order/{id}:
    get:
      tags:
      - Order
      summary: Buscar una orden en específico
      operationId: GetOrder
      consumes: []
      produces:
      - application/json
      parameters:
      - name: id
        in: path
        description: ID de la órden
        required: true
        type: string
      responses:
        200:
          description: Success
          schema:
            $ref: '#/definitions/OrderResponse'
    put:
      tags:
      - Order
      operationId: Put
      consumes:
      - application/json-patch+json
      - application/json
      - text/json
      - application/*+json
      produces:
      - application/json
      parameters:
      - name: id
        in: path
        required: true
        type: string
      - name: model
        in: body
        schema:
          $ref: '#/definitions/OrderEntity'
      responses:
        200:
          description: Success
    delete:
      tags:
      - Order
      operationId: Delete
      consumes: []
      produces:
      - application/json
      parameters:
      - name: id
        in: path
        required: true
        type: string
      responses:
        200:
          description: Success
  /DEV/api/v1/Order/UpdateStatus/{id}:
    put:
      tags:
      - Order
      summary: Actualiza el estado de la orden a "recibido"
      operationId: UpdateStatus
      consumes: []
      produces:
      - application/json
      parameters:
      - name: id
        in: path
        description: ''
        required: true
        type: string
      responses:
        200:
          description: Success
  /DEV/api/v1/Order/UpdateOrderDetails/{idpedido}:
    put:
      tags:
      - Order
      summary: Actualiza el detalle de la orden, el estado del comprobante y actualiza el stock o kardex.
      operationId: UpdateOrderDetails
      consumes:
      - application/json-patch+json
      - application/json
      - text/json
      - application/*+json
      produces:
      - application/json
      parameters:
      - name: idpedido
        in: path
        description: ID del pedido
        required: true
        type: string
      - name: model
        in: body
        description: Detalles del pedido
        schema:
          uniqueItems: false
          type: array
          items:
            $ref: '#/definitions/OrderDetailRequest'
      responses:
        200:
          description: Success
...
  /DEV/api/v1/Product/{id}:
    get:
      tags:
      - Product
      operationId: Get
      consumes: []
      produces:
      - application/json
      parameters:
      - name: id
        in: path
        required: true
        type: string
      responses:
        200:
          description: Success
          schema:
            $ref: '#/definitions/ProductEntity'
    put:
      tags:
      - Product
      operationId: Put
      consumes:
      - application/json-patch+json
      - application/json
      - text/json
      - application/*+json
      produces:
      - application/json
      parameters:
      - name: id
        in: path
        required: true
        type: string
      - name: model
        in: body
        schema:
          $ref: '#/definitions/ProductRequest'
      responses:
        200:
          description: Success
    delete:
      tags:
      - Product
      operationId: Delete
      consumes: []
      produces:
      - application/json
      parameters:
      - name: id
        in: path
        required: true
        type: string
      responses:
        200:
          description: Success
...
  /DEV/api/v1/Provider/{id}:
    get:
      tags:
      - Provider
      operationId: Get
      consumes: []
      produces:
      - application/json
      parameters:
      - name: id
        in: path
        required: true
        type: string
      responses:
        200:
          description: Success
          schema:
            $ref: '#/definitions/ProviderEntity'
    put:
      tags:
      - Provider
      operationId: Put
      consumes:
      - application/json-patch+json
      - application/json
      - text/json
      - application/*+json
      produces:
      - application/json
      parameters:
      - name: id
        in: path
        required: true
        type: string
      - name: model
        in: body
        schema:
          $ref: '#/definitions/ProviderRequest'
      responses:
        200:
          description: Success
    delete:
      tags:
      - Provider
      operationId: Delete
      consumes: []
      produces:
      - application/json
      parameters:
      - name: id
        in: path
        required: true
        type: string
      responses:
        200:
          description: Success
...
  /DEV/api/v1/Receipt/{id}:
    get:
      tags:
      - Receipt
      operationId: GetReceipt
      consumes: []
      produces:
      - application/json
      parameters:
      - name: id
        in: path
        required: true
        type: string
      responses:
        200:
          description: Success
          schema:
            $ref: '#/definitions/ReceiptResponse'
    delete:
      tags:
      - Receipt
      operationId: Delete
      consumes: []
      produces:
      - application/json
      parameters:
      - name: id
        in: path
        required: true
        type: string
      responses:
        200:
          description: Success
...
definitions:
...
  BrandEntity:
    description: Entidad de marcas
    type: object
    properties:
      internalId:
        type: string
      id:
        type: string
      nombre:
        type: string
...
  ProductEntity:
    type: object
    properties:
      internalId:
        type: string
      id:
        type: string
      numero:
        format: int32
        type: integer
      codigo:
        description: Correlativo (P-0001)
        type: string
      nombre:
        type: string
      precioUnitario:
        format: double
        type: number
      moneda:
        enum:
        - Soles
        - Dolares
        type: string
      um:
        enum:
        - Unidad
        - Miligramo
        - Gramo
        - Kilogramo
        - Litro
        type: string
      peso:
        format: double
        type: number
      tipo:
        type: string
      descripcion:
        type: string
      idProveedor:
        type: string
      marca:
        $ref: '#/definitions/BrandEntity'
      updatedOn:
        format: date-time
        type: string
      createdOn:
        format: date-time
        type: string
  OrderResponse:
    type: object
    properties:
      id:
        type: string
      numero:
        format: int32
        type: integer
      codigo:
        type: string
      total:
        format: double
        type: number
      estado:
        type: string
      idUsuario:
        type: string
      numeroFactura:
        type: string
      validado:
        type: boolean
      detallePedido:
        uniqueItems: false
        type: array
        items:
          $ref: '#/definitions/OrderDetailResponse'
      updatedOn:
        format: date-time
        type: string
      createdOn:
        format: date-time
        type: string
  OrderDetailResponse:
    type: object
    properties:
      id:
        type: string
      producto:
        $ref: '#/definitions/ProductEntity'
      cantidad:
        format: int32
        type: integer
      subTotal:
        format: double
        type: number
...
  OrderDetailRequest:
    required:
    - idProducto
    type: object
    properties:
      idProducto:
        type: string
      cantidad:
        format: int32
        maximum: 2147483647
        minimum: 0
        type: integer
  OrderEntity:
    description: Pedidos
    type: object
    properties:
      internalId:
        type: string
      id:
        type: string
      numero:
        format: int32
        type: integer
      codigo:
        description: Correlativo (O-0001)
        type: string
      total:
        format: double
        type: number
      estado:
        description: 'Estado del pedido (default: pendiente)'
        enum:
        - Pendiente
        - Recibido
        type: string
      idUsuario:
        type: string
      idRecibo:
        type: string
      validado:
        description: Campo usado para confirmar si la orden ya se validó en conjunto con el comprobante.
        type: boolean
      detallePedido:
        uniqueItems: false
        type: array
        items:
          $ref: '#/definitions/OrderDetailEntity'
      updatedOn:
        format: date-time
        type: string
      createdOn:
        format: date-time
        type: string
  OrderDetailEntity:
    type: object
    properties:
      internalId:
        type: string
      id:
        type: string
      idProducto:
        type: string
      cantidad:
        format: int32
        type: integer
      subTotal:
        format: double
        type: number
...
  ProviderEntity:
    type: object
    properties:
      internalId:
        type: string
      id:
        type: string
      ruc:
        type: string
      razonSocial:
        type: string
      direccion:
        type: string
      telefono:
        type: string
      contacto:
        type: string
      celular:
        type: string
      email:
        type: string
      descripcion:
        type: string
      updatedOn:
        format: date-time
        type: string
      createdOn:
        format: date-time
        type: string
  ProductRequest:
    required:
    - nombre
    type: object
    properties:
      idProveedor:
        type: string
      nombre:
        type: string
      precioUnitario:
        format: double
        minimum: 0
        type: number
      moneda:
        enum:
        - Soles
        - Dolares
        type: string
      um:
        enum:
        - Unidad
        - Miligramo
        - Gramo
        - Kilogramo
        - Litro
        type: string
      idMarca:
        type: string
      peso:
        format: double
        type: number
      tipo:
        type: string
      descripcion:
        type: string
  ProviderRequest:
    required:
    - ruc
    - razonSocial
    type: object
    properties:
      ruc:
        type: string
      razonSocial:
        type: string
      direccion:
        type: string
      telefono:
        maxLength: 10
        type: string
      contacto:
        maxLength: 50
        type: string
      celular:
        maxLength: 9
        type: string
      email:
        type: string
      descripcion:
        maxLength: 100
        type: string
  ReceiptResponse:
    type: object
    properties:
      id:
        type: string
      numeroFactura:
        type: string
      fechaFactura:
        format: date-time
        type: string
      tipo:
        type: string
      estado:
        type: string
      tieneMedioPago:
        description: >-
          Sólo es usado para mostrar si el comprobante a pesar de estar con estado conforme,

          tiene el valor de tipo de pago en null
        type: boolean
      pedido:
        $ref: '#/definitions/OrderResponse'
      updatedOn:
        format: date-time
        type: string
      createdOn:
        format: date-time
        type: string
...
securityDefinitions:
  Bearer:
    name: Authorization
    in: header
    type: apiKey
    description: 'JWT Authorization header using the Bearer scheme. Example: "Authorization: Bearer {token}"'
security:
- Bearer: []